# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Islamic Andolon Project API",
        "description": "API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile",
                                    "password"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "string",
                                        "example": "01712345678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "mobile",
                                    "password",
                                    "password_confirmation",
                                    "seat_id",
                                    "division_id",
                                    "district_id",
                                    "upazila_id",
                                    "union_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "01712345678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "seat_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "division_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "district_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "upazila_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "union_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Optional message"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully registered"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get Current User",
                "operationId": "f0b15943f55143d36f606a2dbd1d2366",
                "responses": {
                    "200": {
                        "description": "User details"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh Token",
                "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
                "responses": {
                    "200": {
                        "description": "New Token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/districts": {
            "get": {
                "tags": [
                    "Geo"
                ],
                "summary": "Get districts",
                "operationId": "f3578d6f47e9db1b946d9d9ab4711eb2",
                "parameters": [
                    {
                        "name": "pid",
                        "in": "query",
                        "description": "Division ID",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of districts"
                    }
                }
            }
        },
        "/api/upazilas": {
            "get": {
                "tags": [
                    "Geo"
                ],
                "summary": "Get upazilas",
                "operationId": "7ab0547e76958fd85eed02e75bff6947",
                "parameters": [
                    {
                        "name": "pid",
                        "in": "query",
                        "description": "District ID",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of upazilas"
                    }
                }
            }
        },
        "/api/unions": {
            "get": {
                "tags": [
                    "Geo"
                ],
                "summary": "Get unions",
                "operationId": "f5543f67a77c44c7913084bf6e29b47d",
                "parameters": [
                    {
                        "name": "pid",
                        "in": "query",
                        "description": "Upazila ID",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of unions"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}